Create matrices
Create a program that gets the number of test cases or matrices (n) from the standard
input. Then, from the standard input, the program also gets the information needed to
generate random matrices with different shapes. Then write these matrices to a file
output.txt. It then calculates the determinant of each matrix and writes this to the same
file under each corresponding matrices.